Add multi-gNB support in l2proxy, channel simulation, and tutorial scripts
- Multi-gNB L2 proxy with P5/P7 connection support for multiple gNBs
- gNB/UE configuration files for L2 proxy mode
- Channel simulation with per-slot SINR and HARQ error modeling
- Channel trace file reader thread and trace generation utility
- ClickHouse DB integration for PM metrics publishing
- UL DCI and UL TTI request filtering by RNTI
- Multi-threaded downlink message processing for improved throughput
- Non-blocking queue with memory pool pattern for DL UE messages
- pf_ul scheduler prioritization for blocked users
- Cell selection parameters stored from SIB in UE RRC state
- Signal handler for graceful shutdown
- Docker dev container updates (gNB, base, build)
- Tutorial scripts for multi-gNB UE simulation (Kubernetes deployment,
  iperf traffic, UE lifecycle management)
- Run script for L2 proxy with dynamic gNB/UE count
- Stability improvements: CRC/RX data indication synchronization,
  TX data request queue handling, slot time optimization

+++ openair2/COMMON/mac_messages_types.h
@@ -120,7 +120,7 @@ typedef struct NRRrcMacBcchDataInd_s {
   uint8_t   slot;
   uint32_t  sdu_size;
   uint8_t   sdu[BCCH_SDU_SIZE];
-  uint8_t   gnb_index;
+  uint16_t   gnb_index;
   bool      is_bch;
   uint8_t   rsrq;
   uint8_t   rsrp;
